+++ plugins/node/opentelemetry-instrumentation-mongodb/test/mongodb-v3.test.ts
@@ -112,7 +112,12 @@ describe('MongoDBInstrumentation', () => {
           .insertMany(insertData)
           .then(() => {
             span.end();
-            assertSpans(getTestSpans(), 'mongodb.insert', SpanKind.CLIENT);
+            assertSpans(
+              getTestSpans(),
+              'mongodb.insert',
+              SpanKind.CLIENT,
+              'insert'
+            );
             done();
           })
           .catch(err => {
@@ -128,7 +133,12 @@ describe('MongoDBInstrumentation', () => {
           .updateOne({ a: 2 }, { $set: { b: 1 } })
           .then(() => {
             span.end();
-            assertSpans(getTestSpans(), 'mongodb.update', SpanKind.CLIENT);
+            assertSpans(
+              getTestSpans(),
+              'mongodb.update',
+              SpanKind.CLIENT,
+              'update'
+            );
             done();
           })
           .catch(err => {
@@ -144,7 +154,12 @@ describe('MongoDBInstrumentation', () => {
           .deleteOne({ a: 3 })
           .then(() => {
             span.end();
-            assertSpans(getTestSpans(), 'mongodb.remove', SpanKind.CLIENT);
+            assertSpans(
+              getTestSpans(),
+              'mongodb.remove',
+              SpanKind.CLIENT,
+              'remove'
+            );
             done();
           })
           .catch(err => {
@@ -164,7 +179,12 @@ describe('MongoDBInstrumentation', () => {
           .toArray()
           .then(() => {
             span.end();
-            assertSpans(getTestSpans(), 'mongodb.find', SpanKind.CLIENT);
+            assertSpans(
+              getTestSpans(),
+              'mongodb.find',
+              SpanKind.CLIENT,
+              'find'
+            );
             done();
           })
           .catch(err => {
@@ -190,7 +210,8 @@ describe('MongoDBInstrumentation', () => {
                   span => !span.name.includes('mongodb.getMore')
                 ),
                 'mongodb.find',
-                SpanKind.CLIENT
+                SpanKind.CLIENT,
+                'find'
               );
               // assert that we correctly got the first as a find
               assertSpans(
@@ -198,7 +219,8 @@ describe('MongoDBInstrumentation', () => {
                   span => !span.name.includes('mongodb.find')
                 ),
                 'mongodb.getMore',
-                SpanKind.CLIENT
+                SpanKind.CLIENT,
+                'getMore'
               );
               done();
             })
@@ -222,7 +244,8 @@ describe('MongoDBInstrumentation', () => {
             assertSpans(
               getTestSpans(),
               'mongodb.createIndexes',
-              SpanKind.CLIENT
+              SpanKind.CLIENT,
+              'createIndexes'
             );
             done();
           })
@@ -253,7 +276,14 @@ describe('MongoDBInstrumentation', () => {
             span.end();
             const spans = getTestSpans();
             const operationName = 'mongodb.insert';
-            assertSpans(spans, operationName, SpanKind.CLIENT, false, false);
+            assertSpans(
+              spans,
+              operationName,
+              SpanKind.CLIENT,
+              'insert',
+              false,
+              false
+            );
             const mongoSpan = spans.find(s => s.name === operationName);
             const dbStatement = JSON.parse(
               mongoSpan!.attributes[SemanticAttributes.DB_STATEMENT] as string
@@ -291,7 +321,14 @@ describe('MongoDBInstrumentation', () => {
               span.end();
               const spans = getTestSpans();
               const operationName = 'mongodb.insert';
-              assertSpans(spans, operationName, SpanKind.CLIENT, false, true);
+              assertSpans(
+                spans,
+                operationName,
+                SpanKind.CLIENT,
+                'insert',
+                false,
+                true
+              );
               const mongoSpan = spans.find(s => s.name === operationName);
               const dbStatement = JSON.parse(
                 mongoSpan!.attributes[SemanticAttributes.DB_STATEMENT] as string
@@ -324,7 +361,7 @@ describe('MongoDBInstrumentation', () => {
             .then(() => {
               span.end();
               const spans = getTestSpans();
-              assertSpans(spans, 'mongodb.insert', SpanKind.CLIENT);
+              assertSpans(spans, 'mongodb.insert', SpanKind.CLIENT, 'insert');
               done();
             })
             .catch(err => {
@@ -421,7 +458,7 @@ describe('MongoDBInstrumentation', () => {
             .then(() => {
               span.end();
               const spans = getTestSpans();
-              assertSpans(spans, 'mongodb.find', SpanKind.CLIENT);
+              assertSpans(spans, 'mongodb.find', SpanKind.CLIENT, 'find');
               done();
             })
             .catch(err => {
@@ -443,7 +480,7 @@ describe('MongoDBInstrumentation', () => {
             span.end();
             const spans = getTestSpans();
             const mainSpan = spans[spans.length - 1];
-            assertSpans(spans, 'mongodb.insert', SpanKind.CLIENT);
+            assertSpans(spans, 'mongodb.insert', SpanKind.CLIENT, 'insert');
             resetMemoryExporter();
 
             collection
@@ -453,7 +490,7 @@ describe('MongoDBInstrumentation', () => {
                 const spans2 = getTestSpans();
                 spans2.push(mainSpan);
 
-                assertSpans(spans2, 'mongodb.find', SpanKind.CLIENT);
+                assertSpans(spans2, 'mongodb.find', SpanKind.CLIENT, 'find');
                 assert.strictEqual(
                   mainSpan.spanContext().spanId,
                   spans2[0].parentSpanId
@@ -510,11 +547,11 @@ describe('MongoDBInstrumentation', () => {
             (err, address) => {
               if (err) return done(err);
               assert.strictEqual(
-                mongoSpan.attributes[SemanticAttributes.NET_HOST_NAME],
+                mongoSpan.attributes[SemanticAttributes.NET_PEER_NAME],
                 address
               );
               assert.strictEqual(
-                mongoSpan.attributes[SemanticAttributes.NET_HOST_PORT],
+                mongoSpan.attributes[SemanticAttributes.NET_PEER_PORT],
                 process.env.MONGODB_PORT || '27017'
               );
               done();