+++ plugins/node/opentelemetry-instrumentation-mysql2/src/index.ts
@@ -19,4 +19,4 @@ import { MySQL2Instrumentation } from './instrumentation';
 export * from './instrumentation';
 export default MySQL2Instrumentation;
 
-export { MySQL2InstrumentationConfig } from './types';
+export * from './types';