+++ plugins/node/opentelemetry-instrumentation-net/test/utils.ts
@@ -58,6 +58,7 @@ export function assertTLSSpan(
   { netSpan, tlsSpan }: { netSpan: ReadableSpan; tlsSpan: ReadableSpan },
   socket: Socket
 ) {
+  assertParentChild(tlsSpan, netSpan);
   assertSpanKind(netSpan);
   assertAttrib(
     netSpan,
@@ -106,6 +107,17 @@ export function assertSpanKind(span: ReadableSpan) {
 
 export function assertAttrib(span: ReadableSpan, attrib: string, value: any) {
   assert.strictEqual(span.attributes[attrib], value);
+}
+
+export function assertParentChild(
+  parentSpan: ReadableSpan,
+  childSpan: ReadableSpan
+) {
+  assert.strictEqual(
+    childSpan.spanContext().traceId,
+    parentSpan.spanContext().traceId
+  );
+  assert.strictEqual(childSpan.parentSpanId, parentSpan.spanContext().spanId);
 }
 
 export const TLS_SERVER_CERT = fs